refactor: proposal pool draft
